syntax error fix + gitingore file

+++ bamazon.sql
@@ -9,7 +9,7 @@ CREATE TABLE products (
   product_name VARCHAR(100) NULL,
   department_name VARCHAR(100) NULL,
   price FLOAT(10, 2)  NULL,
-  stock_quantity INTEGER (10000) NULL,
+  stock_quantity INTEGER (8000) NULL,
   PRIMARY KEY (item_id)
 );
